[DependencyInjection] use/fix newest Definition::setFactory
