+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/templating_php.xml
@@ -55,7 +55,8 @@
             <argument /> <!-- version format -->
         </service>
 
-        <service id="templating.asset.request_aware_package" class="Symfony\Component\Templating\Asset\PackageInterface" factory-service="templating.asset.package_factory" factory-method="getPackage" abstract="true">
+        <service id="templating.asset.request_aware_package" class="Symfony\Component\Templating\Asset\PackageInterface" abstract="true">
+            <factory service="templating.asset.package_factory" method="getPackage" />
             <argument type="service" id="request" strict="false" />
             <argument /> <!-- HTTP id -->
             <argument /> <!-- SSL id -->
