+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/validator.xml
@@ -16,9 +16,12 @@
     </parameters>
 
     <services>
-        <service id="validator" class="%validator.class%" factory-service="validator.builder" factory-method="getValidator" />
+        <service id="validator" class="%validator.class%">
+            <factory service="validator.builder" method="getValidator" />
+        </service>
 
-        <service id="validator.builder" class="%validator.builder.class%" factory-class="%validator.builder.factory.class%" factory-method="createValidatorBuilder">
+        <service id="validator.builder" class="%validator.builder.class%">
+            <factory class="%validator.builder.factory.class%" method="createValidatorBuilder" />
             <call method="setConstraintValidatorFactory">
                 <argument type="service" id="validator.validator_factory" />
             </call>
